Time: 0 ms (100%), Space: 40.7 MB (80%) - LeetHub

+++ 2837-minimum-operations-to-make-the-integer-zero/2837-minimum-operations-to-make-the-integer-zero.java
@@ -8,7 +8,7 @@ class Solution {
             }
 
             if(Long.bitCount(val) <= i && i <= val) {
-                return (int)i;
+                return i;
             }
         }
     }
